% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOSM.R
\name{cyclegraphOSM}
\alias{cyclegraphOSM}
\title{cyclegraphOSM}
\usage{
cyclegraphOSM(input, savename, inc_primary = FALSE,
  inc_secondary = TRUE)
}
\arguments{
\item{input}{The input file path}

\item{savename}{Output name}

\item{inc_primary}{If TRUE, primary roads are included. Default is FALSE}

\item{inc_secondary}{If TRUE, secondary roads are included. Default is TRUE}
}
\value{
The osmosis call
}
\description{
Create a cyclist routing graph.
}
\examples{
\donttest{
dest <- getOSM(filterby="osm", exclude = "md5", r1 = 2, r2 =13, dest="")
cyclegraphOSM(dest)
}
}
\author{
Sebastian Gatscha
}
