% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOSM.R
\name{filterSF}
\alias{filterSF}
\title{filterSF}
\usage{
filterSF(sf, tf = NULL, id = NULL, version = NULL, nodes = NULL,
  timestamp = NULL, plot = FALSE, trimtags = TRUE, pal, ...)
}
\arguments{
\item{sf}{Input Simple Feature}

\item{tf}{Filter by tag key/values}

\item{id}{Filtery by id}

\item{version}{Filtery by version}

\item{nodes}{Filtery by nodes (Only for ways)}

\item{timestamp}{Filtery by timestamp. The column will be detected, if
it is configured as POSIXct.}

\item{plot}{Plot the result? Default is FALSE.}

\item{trimtags}{Trim teh tag names? Default is TRUE}

\item{pal}{A color palette}

\item{...}{Passed on to \code{\link{grep}}}
}
\value{
The filtered Simple Feature
}
\description{
Filter the Simple Feature output from \code{\link{sfOSM}}.
}
\examples{
\donttest{
dblist <- list(dbname="test", dbuser="postgres", dbhost="localhost",
               dbport="5432", dbpwd="hallo")

## Working with OSM Ways
ways <- sfOSM(dblist, elem="ways")
r <- filterSF(ways, id=4383033, plot=T);
r <- filterSF(ways, nodes="5339632222,5343805366,5343805367,5343805368,5343805369,5339632232",
              plot=T, fixed=T);
r <- filterSF(ways, tf="primary", version = 11, plot=T);
r <- filterSF(ways, tf=c("footway", "path"), version = c(11,12), plot=T);
r <- filterSF(ways, id=184691002, plot=T);
r <- filterSF(ways, tf="surface\\"=>\\"asphalt", plot=T);
r <- filterSF(ways, tf=c("footway", "path"), plot=T, pal=c("blue","red"));
r <- filterSF(ways, tf=c("secondary", "residential","footway", "path"),
              plot=T, pal=rainbow(5));

## Working with OSM Nodes
nodes <- sfOSM(dblist, elem="nodes")
r <- filterSF(nodes, tf="highway", plot=T);
r <- filterSF(nodes, tf="crossing", version=11, plot=T);
r <- filterSF(nodes, version=9,  plot=T);
r <- filterSF(nodes, id=26936524,  plot=T);
}
}
\author{
Sebastian Gatscha
}
