% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOSM.R
\name{getOSM}
\alias{getOSM}
\title{getOSM}
\usage{
getOSM(filterby = NULL, exclude = NULL, r1 = NULL, r2 = NULL,
  r3 = NULL, dest = NULL)
}
\arguments{
\item{filterby}{Filter by extensions}

\item{exclude}{Exclude by extensions}

\item{r1}{First link (Continent)}

\item{r2}{Second link (Countries)}

\item{r3}{Third link (Download / Sub Regions)}

\item{dest}{A destination directory}
}
\value{
The download destination
}
\description{
Get OSM-Data from
\href{https://www.geofabrik.de/}{Geofabrik GmbH}
}
\examples{
\donttest{
dest <- getOSM()
dest <- getOSM(exclude = "md5", r1 = 2, r2=13)
dest <- getOSM(filterby="shp", r1 = 2, r2=13)
dest <- getOSM(filterby="osm", exclude = "md5", r1 = 2, r2 =13, dest="")
}
}
\author{
Sebastian Gatscha
}
