% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOSM.R
\name{graphcarOSM}
\alias{graphcarOSM}
\title{graphcarOSM}
\usage{
graphcarOSM(input, savename)
}
\arguments{
\item{input}{The input file path}

\item{savename}{Output name}
}
\value{
The osmosis call
}
\description{
Create a car routing graph.
}
\examples{
\donttest{
dest <- getOSM(filterby="osm", exclude = "md5", r1 = 2, r2 =13, dest="")
graphcarOSM(dest)
}
}
\author{
Sebastian Gatscha
}
