% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOSM.R
\name{mapviewOSM}
\alias{mapviewOSM}
\title{mapviewOSM}
\arguments{
\item{sumry}{The output of summaryOSM}

\item{mergeby}{Should the values be assigned per continent
or per country. Default is 'country'}

\item{unit}{Which unit should be taken. Default is 'mb'}

\item{...}{Arguments passed on to mapview}
}
\description{
Plot the OSM summary dataset with mapview.
}
\examples{
\donttest{
con_df <- summaryOSM()
mapviewOSM(con_df, mergeby = "subregions", unit = "gb")
}
}
\author{
Sebastian Gatscha
}
