% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOSM.R
\name{osmosisR}
\alias{osmosisR}
\title{osmosisR}
\usage{
osmosisR(input, savename = "output.osm", idtracker = FALSE,
  usednode = FALSE, usedway = FALSE, extract_nodes = NULL,
  extract_ways = NULL, bbox = NULL, poly = NULL,
  filter_nodes = NULL, nodes_accept = TRUE, filter_ways = NULL,
  ways_accept = TRUE, filter_relations = NULL,
  relation_accept = TRUE)
}
\arguments{
\item{input}{The input file path}

\item{savename}{Output name}

\item{idtracker}{If TRUE, sets idTrackerType=BitSet}

\item{usednode}{If TRUE, sets --used-node}

\item{usedway}{If TRUE, sets --used-way}

\item{extract_nodes}{Must be given as key-value pair, like 'natural.tree'}

\item{extract_ways}{Must be given as key-value pair.}

\item{bbox}{Extract data with bounding box coordinates in lat/lon}

\item{poly}{Extract data with .POLY files}

\item{filter_nodes}{Use --tag-filter with nodes}

\item{nodes_accept}{logical, if nodes should be accepted or rejected.
Default is TRUE}

\item{filter_ways}{Use --tag-filter with ways}

\item{ways_accept}{logical, if ways should be accepted or rejected.
Default is TRUE}

\item{filter_relations}{Use --tag-filter with relations}

\item{relation_accept}{logical, if relations should be accepted or
rejected. Default is TRUE}
}
\value{
The osmosis call.
}
\description{
Basic Interface to osmosis.
}
\examples{
\donttest{
dest <- getOSM(filterby="osm", exclude = "md5", r1 = 2, r2 =13, dest="")
osmosisR(dest, savename = "highway_capeverde.osm",
         filter_ways = "highway=*")

osmosisR(dest,
         savename = "motorways_dest.osm",
         usednode = T, usedway = T,
         ways_accept = F,filter_ways = "",
         filter_nodes = "natural=tree")

## Get only tags with 'building=yes' and reject ways and relations.
osmosisR(dest,
         savename = "buildings_dest.osm",
         filter_relations = T,filter_ways = "",
         filter_nodes = "building=yes",
         relation_accept = F, nodes_accept =  T,
         ways_accept = F)

## Filter for multiple tag values (beach, tree)
osmosisR(dest,
        savename = "buildings_dest.osm",
        filter_nodes = "natural=beach,tree")

## Extract data based on node key-values
osmosisR(dest,
         savename = "tree_beach_dest.osm",
         usednode=F, usedway = T,
         extract_nodes = "natural.tree,natural.beach")

}
}
\author{
Sebastian Gatscha
}
