% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOSM.R
\name{postgresOSM}
\alias{postgresOSM}
\title{postgresOSM}
\usage{
postgresOSM(input, dblist)
}
\arguments{
\item{input}{The input file path (Must be in .osm or .pbf format)}

\item{dblist}{A list with database credentials.}
}
\value{
The osmosis call
}
\description{
Import OSM/PBF file to PostgreSQL.
}
\examples{
\donttest{
dest <- getOSM(filterby="osm", exclude = "md5", r1 = 2, r2 =13, dest="")
dblist <- list(dbname="test", dbuser="postgres", dbhost="localhost",
               dbport="5432", dbpwd="postgres")
cmd <- postgresOSM(dest, dblist)
}
}
\author{
Sebastian Gatscha
}
