% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOSM.R
\name{sfOSM}
\alias{sfOSM}
\title{sfOSM}
\usage{
sfOSM(dblist, elem = "ways", crsin = 4326, crsout = NULL)
}
\arguments{
\item{dblist}{A list with database credentials.}

\item{elem}{Which elements should be converted to Simple Features?
Must be one of 'nodes' or 'ways'. Default is 'ways'.}

\item{crsin}{Input CRS. Default is 4326}

\item{crsout}{Output CRS, if transformation is needed. Default is NULL}
}
\value{
A Simple Feature object with either POINTS or LINESTRINGS
}
\description{
Read nodes or ways from PostgreSQL to R as sf.
}
\examples{
\donttest{
dblist <- list(dbname="test", dbuser="postgres", dbhost="localhost",
              dbport="5432", dbpwd="hallo")
ways <- sfOSM(dblist)
plot(st_geometry(ways))

nodes <- sfOSM(dblist, elem = "nodes")
plot(st_geometry(nodes))

ways3035 <- sfOSM(dblist, crsout=3035)
plot(st_geometry(ways3035))
}
}
\author{
Sebastian Gatscha
}
