% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOSM.R
\name{summaryOSM}
\alias{summaryOSM}
\title{summaryOSM}
\usage{
summaryOSM()
}
\value{
A summary dataset, which can be used for treemapOSM or mapviewOSM.
}
\description{
Get a summary dataset of OSM
}
\examples{
\donttest{
con_df <- summaryOSM()
}
}
\author{
Sebastian Gatscha
}
