% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOSM.R
\name{treemapOSM}
\alias{treemapOSM}
\title{treemapOSM}
\usage{
treemapOSM(sumry)
}
\arguments{
\item{sumry}{The output of summaryOSM}
}
\description{
Plot a treemap of the OSM summary dataset
}
\examples{
\donttest{
con_df <- summaryOSM()
treemapOSM(con_df)
}
}
\author{
Sebastian Gatscha
}
