% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOSM.R
\name{convertOSM}
\alias{convertOSM}
\title{convertOSM}
\usage{
convertOSM(destfin, ext = "osm", cm = FALSE, cb = FALSE,
  cw = FALSE, bbox = NULL, fname = NULL, poly = NULL)
}
\arguments{
\item{destfin}{The input file path.}

\item{ext}{Possible extensions.}

\item{cm}{Include --complete-multipolygons}

\item{cb}{Include --complete-boundaries}

\item{cw}{Include --complete-ways}

\item{bbox}{Cut out the result via bbox.}

\item{fname}{Output file name}

\item{poly}{Cut out the result via a .poly-file}
}
\value{
The osmconvert call.
}
\description{
Interface to osmconvert. (It must be set as system variable)
}
\examples{
\donttest{
dest <- getOSM(filterby="osm", exclude = "md5", r1 = 2, r2 =13, dest="")
convertOSM(dest, cm=T, cb = F, cw = F, fname = "capverde4", ext = "pbf")
}
}
\author{
Sebastian Gatscha
}
