% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOSM.R
\name{mapviewOSM}
\alias{mapviewOSM}
\title{mapviewOSM}
\arguments{
\item{sumry}{The output of summaryOSM}

\item{mergeby}{Should the values be assigned per continent
or per country. Default is 'country'}

\item{unit}{Which unit should be taken. Default is 'mb'}

\item{...}{Arguments passed on to mapview.
See \href{https://r-spatial.github.io/mapview}{Mapview}}
}
\description{
Plot the OSM summary dataset with mapview.
}
\examples{
\donttest{
## Get a summary dataset and plot it
con_df <- summaryOSM()
mapviewOSM(con_df, mergeby = "country", unit = "gb")

## With Custom col.regions
YlOrBr <- c("#FFFFD4", "#FED98E", "#FE9929", "#D95F0E", "#993404")
colfunc = colorRampPalette(YlOrBr, space = "Lab")
mapviewOSM(con_df, mergeby = "country", col.regions = colfunc)

## With custon col.regions and breaks
jet.colors <- colorRampPalette(c("#00007F", "blue", "#007FFF", "cyan",
                       "#7FFF7F", "yellow", "#FF7F00", "red", "#7F0000"))
mapviewOSM(con_df, mergeby = "country", unit = "gb", burst=F,
          col.regions = jet.colors, at = 0:3)

## With custom popups
mapviewOSM(con_df, mergeby = "country",
           popup = mapview::popupTable(con_df, zcol = c("subregions", "kilobyte"),
                                      feature.id = F))
}
}
\author{
Sebastian Gatscha
}
